doc-helper/pyproject.toml
[project]
name = "doc-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "black>=25.1.0",
    "fastapi>=0.115.12",
    "firecrawl-py>=1.16.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-openai>=0.3.13",
    "langchain-pinecone>=0.2.5",
    "nltk>=3.9.1",
    "openai>=1.75.0",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
    "streamlit-chat>=0.1.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "unstructured>=0.17.2",
    "uvicorn>=0.34.1",
]
